+++ providers/utils/index.js
@@ -1,0 +1,3 @@
+const { getCountryAlpha2Code } = require("./countries");
+
+module.exports = { getCountryAlpha2Code };